Tag Alpine containers with 3.17, not 3.16
The container image uses the Alpine version that is provided by the
Elclipse Temurin Alpine container image.  That container image was
upgraded from 3.16 to 3.17 months ago.  We didn't detect that version
upgrade. We don't have a test that checks the Alpine version matches
the tag we are creating on Dockerhub.

This is a short term solution, since it doesn't add any check to verify
the Alpine version included in the Eclipse Temurin Alpine container
matches the version we use to create the Dockerhub tag.

Since Alpine 3.18.0 has released, we can expect that the next Eclipse
Temurin update (11.0.20, 17.0.8) in about 3 months will deliver Alpine
3.18.0 instead of 3.17.3.

Fixes https://github.com/jenkinsci/docker-agent/issues/415