fix(alpine) The Alpine Dockerfile does not use the `ALPINE_TAG` arg to build
We are getting the Alpine version shipped with Eclipse Temurin.
We don't take into account the `ALPINE_TAG` set in the `docker-bake.hcl` file.
Damien's proposal was to do as we do with Debian, with two stages, one to get the Temurin build, and one to get the correct (i.e. defined in the `docker-bake`) Alpine version.
The (optional?) next step would be to use `jlink` as in the Debian image.