[JENKINS-73703] Remove UBI 8 + Review Feedback
With the upcoming EOL of Java 11, the UBI 8 image for Jenkins server
will no longer be supported. Creating a UBI 8 image therefore does not
make any sense.

The tag for the base UBI 9 image was converted to a build arg, to
facilitate automatic updates via updatecli. Other minor fixes to
documentation and image metadata also included.

Signed-off-by: Adam Kaplan <adam.kaplan@redhat.com>
Co-authored-by: Mark Waite <mark.earl.waite@gmail.com>
Co-authored-by: lemeurherveCB <137290663+lemeurherveCB@users.noreply.github.com>