Install `git-lfs` config to system rather than global (#1109) (#1111)
The upstream install script installs the config globally (scoped to the
user).  It is run as root, thus installed to `/root/.gitconfig` and
unavailable to the jenkins user.  Installing to the system, as distro
packages do, makes the config available to any user (i.e. jenkins).

The script currently has no way to change the scope or to opt out of the
 installation.  The global config could be safely removed to save a few
bytes, but opting for a slightly shorter Dockerfile instead.